+++ NEWS.md
@@ -1,5 +1,9 @@
 # gghighlight 0.1.0.9000
 
+* `gghighlight()` gets a new argument `unhighlighted_params`, which accepts a
+  list of parameters for the unhighlighted layer (e.g. `colour`, `fill`, `shape`,
+  and `size`). Accordingly, `unhighlighted_colour` is deprecated (#76).
+
 # gghighlight 0.1.0
 
 * Add `gghighlight()`, which replaces the current `gghighlight_line()` and `gghighlight_point()`; these functions are now deprecated.