+++ NEWS.md
@@ -7,6 +7,9 @@
 * If the mapping has `group`, use it as grouping variable, which is consistent
   with the logic of ggplot2 (#77).
 
+* `gghighlight()` now ignores if the calculation fails over some layers. This
+  is useful to combine with such layers as `annotate()` (#78).
+
 # gghighlight 0.1.0
 
 * Add `gghighlight()`, which replaces the current `gghighlight_line()` and `gghighlight_point()`; these functions are now deprecated.