+++ NEWS.md
@@ -1,4 +1,4 @@
-# gghighlight 0.1.0.9000
+# gghighlight 0.2.0
 
 ## New features